Banyak yang diubah

+++ ATM_D4/src/com/appl/atm/model/Bisnis.java
@@ -4,7 +4,7 @@ public class Bisnis extends Customer {
     private static double MAX_WITHDRAWAL = 2000;
     private static double MAX_TRANSFER   = 10000;
 
-    public Bisnis(int theAccountNumber, int thePIN, int theAvailableBalance, int theTotalBalance) {
+    public Bisnis(int theAccountNumber, int thePIN, double theAvailableBalance, double theTotalBalance) {
         super(theAccountNumber, thePIN, theAvailableBalance, theTotalBalance);
     }
